# specs/001-druids-build-a/contracts/api.yaml
openapi: 3.0.3
info:
  title: Druids Multi-Agent System API
  version: 1.0.0
  description: MCP Server interface for managing agents, scenarios, and coordination
servers:
  - url: http://localhost:3000
    description: Local development server
paths:
...
  /tools/access:
    post:
      summary: Grant tool access to agent
      operationId: grantToolAccess
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ToolAccessGrant'
      responses:
        '201':
          description: Access granted successfully
        '400':
          description: Invalid access grant
        '404':
          description: Agent or tool not found
components:
  schemas:
...
    ToolAccessGrant:
      type: object
      properties:
        agentId:
          type: string
        mcpServerId:
          type: string
        toolName:
          type: string
        permissions:
          type: array
          items:
            type: string
            enum: [invoke, configure]
      required:
        - agentId
        - mcpServerId
        - toolName
        - permissions